==> 1. Data Preparation Meteo.ipynb <==
# Dati Meteo
### Claudia Gabriela Ochoa Martinez (2020)

### Loading Data
#### I dati relativi alle precipitazioni, temperature sono stati reperiti dal Centro Funzionale - Regione autonoma della Valle D'aosta 
##### https://cf.regione.vda.it/ // Scheda stazione meteo "Cogne - Grand Crot"
##### Questi dati sono serviti a rilevare la temperatura e precipitazioni al momento di ogni occasione di controllo delle trappole, i file sono stati salvati sulla directory in formato .CSV dopo essere stati elaborati utilizzando la metodologia Tidy data su Excel

#carichiamo i dati relativi al meteo interessati
files_termometro <- list.files("Dati-originali\\dati_meteo_vda\\", pattern=".*Grand-Crot-Termometro.csv", recursive=TRUE, full.names=TRUE)
files_pluviometro <- list.files("Dati-originali\\dati_meteo_vda\\", pattern=".*Grand-Crot-Pluviometro.csv", recursive=TRUE, full.names=TRUE)
files_vento <- list.files("Dati-originali\\dati_meteo_vda\\", pattern=".*Grand-Crot-Velocita Vento Vett..csv", recursive=TRUE, full.names=TRUE)

precipitazioni_grand_crot <- do.call(rbind, lapply(files_pluviometro, function(x) read.csv(x, dec=",", sep=";", header = TRUE )))
temperatura_grand_crot = do.call(rbind, lapply(files_termometro, function(x) read.csv(x, dec=",", sep=";", header = TRUE )))
velvento_grand_crot = do.call(rbind, lapply(files_vento, function(x) read.csv(x, dec=",", sep=";", header = TRUE )))

##### è stato creato un solo dataframe con relativa data, misurazione precipitazione e temperatura

#precipitazioni_grand_crot$precipitazioni <- lapply(precipitazioni_grand_crot$precipitazioni, function(x) ifelse(is.na(x),"-1",x))
#precipitazioni_grand_crot$precipitazioni <- as.character(precipitazioni_grand_crot$precipitazioni)              
#temp_prec_grand_crot <- merge(precipitazioni_grand_crot,temperatura_grand_crot, by = "Data_Ora_R", all=TRUE)
temp_prec_grand_crot <- cbind(precipitazioni_grand_crot,temperatura_grand_crot$temperatura,velvento_grand_crot$velocita.vento)
tail(temp_prec_grand_crot)


##### Le variabili "Data", sono state trasformate in un formato riconosciuto da R per semplicare future analisi

temp_prec_grand_crot["Orario_R"] <- strftime(as.POSIXct(temp_prec_grand_crot$orario, format="%d/%m/%Y %H:%M", tz = "Europe/Rome"), format="%H:%M")
temp_prec_grand_crot["Data_R"]   <- strftime(as.POSIXct(temp_prec_grand_crot$orario, format="%d/%m/%Y %H:%M", tz = "Europe/Rome"), format="%d-%m-%Y")
temp_prec_grand_crot <- subset( temp_prec_grand_crot, select = -orario)
tail(temp_prec_grand_crot)

colnames(temp_prec_grand_crot) <- c("Precipitazioni","Temperatura.C","velocita.vento","Orario.R","Data.R")
tail(temp_prec_grand_crot)

#### Salviamo il nuovo file .csv nella directory

write.csv2(temp_prec_grand_crot,"Dati-elaborati\\Dati_precip-temp_grand-crot.csv", row.names = FALSE)
