==> 2. Data Preparation.ipynb <==
# Data Preparation
### Claudia Gabriela Ochoa Martinez (2020)
## Loading Data

# CSV Raccolta dati
Daticatt <- read.csv("Dati-originali/daticatture.csv", dec=",", sep=";", header = TRUE)
Datitrap <- read.csv("Dati-originali/datitrappole.csv", dec=",", sep=";", header = TRUE)
Datioccas <- read.csv("Dati-originali/sforzocampionamento.csv", dec=",", sep=";", header = TRUE)
# CSV Meteo
temp_prec_grand_crot <- read.csv("Dati-elaborati\\Dati_precip-temp_grand-crot.csv", dec=",", sep=";", header = TRUE)

### Dati Metereologici

# Convertiamo i dati con un formato riconoscibile da R
temp_prec_grand_crot["Data.R"]   <- as.Date(temp_prec_grand_crot$Data.R, format="%d-%m-%Y")
head(temp_prec_grand_crot)

### Dati Trappole


# Rimuoviamo le righe che non hanno un ID_TRAP - problema in fase di export da Excel
Datitrap <- Datitrap[!is.na(Datitrap$ID_TRAP), ]

# Convertiamo la copertura vegetale in ottavi
Datitrap$Cop_Veg_2015_ottavi <- Datitrap$X..COP..VEG..2015 / 100*8
Datitrap$Cop_Veg_2018_ottavi <- Datitrap$X..COP..VEG..2018 / 100*8

head(Datitrap)

### Dati Sforzo di Campionamento
#### Modificati includendo la data in formato R e l'anno

Datioccas <- Datioccas[!is.na(Datioccas$occasione), ]

Datioccas["Data.R"] <- as.Date(Datioccas$DATA, "%d/%m/%Y")
Datioccas["Anno"] <- format(Datioccas["Data.R"],"%Y")
# Eliminiamo la vecchia colonna Data
Datioccas <- subset( Datioccas, select = -DATA)

# Verifichiamo errori nella colonna relativa all'alba
errori  <- subset(Datioccas,ALBA..A.T. != "A" & ALBA..A.T. != "T")
if (nrow(errori) > 0) {
    print("ATTENZIONE: Valori spuri")
    errori
}

# Aggiungiamo una colonna con l'orario stimato
Datioccas$Orario.R <- lapply(Datioccas$ALBA..A.T., function(x) ifelse(x == "A","07:30","17:30"))
Datioccas$Orario.R <- as.character(Datioccas$Orario.R)

head(Datioccas)

#### Unione dei dati sullo sforzo di campionamento e quelli metereologici

Datioccas <- merge(Datioccas,temp_prec_grand_crot)
head(Datioccas)

### Dati Catture

#### I dati delle catture vanno puliti. Andiamo a rimuovere i campi vuoti ed indeterminati, sostituendoli con NA

Daticatt[Daticatt==""] <- NA
Daticatt[Daticatt=="IND"] <- NA
# Verifichiamo errori nella colonna relativa all'alba

errori  <- subset(Daticatt,ALBA != "A" & ALBA != "T")
if (nrow(errori) > 0) {
    print("ATTENZIONE: Valori spuri")
    errori
}
head(Daticatt)

# Per usi futuri
#Daticatt["Marcatura"] <- paste(Daticatt$chip,Daticatt$pelo)

#### Trasformiamo la Data in formato comprensibile da R ed aggiungiamo una colonna per semplificare le analisi di anno in anno

Daticatt["Data_R"] <- as.Date(Daticatt$DATA, "%d/%m/%Y")
Daticatt["Anno"] <- format(Daticatt["Data_R"],"%Y")
# Eliminiamo la vecchia colonna Data
Daticatt <- subset( Daticatt, select = -DATA)

head(Daticatt)

#### Per alcune analisi, risulta utile unire il dataframe delle catture con quello delle trappole

mergedData <- merge(Daticatt, Datitrap, by.x=c("ID_TRAPPOLA"), by.y=c("ID_TRAP"))

head(mergedData) 

### Sono stati creati e salvati i nuovi dataframe da utilizzare per le analisi

write.csv2(Daticatt,"Dati-elaborati\\Dati-catture.csv", row.names = FALSE)
write.csv2(Datitrap,"Dati-elaborati\\Dati-trappole.csv", row.names = FALSE)
write.csv2(Datioccas,"Dati-elaborati\\Dati-occasioni.csv", row.names = FALSE)
write.csv2(mergedData,"Dati-elaborati\\Dati-catture-trappole.csv", row.names = FALSE)

### creiamo anche un dataframe che includono le covariate
dati.cov <- Datitrap[,c("Diametro.roccia.1..cm.","Diametro.roccia.2..cm.","Diametro.roccia.3..cm.","alt_m","aspect_deg","slope_degr","n_species","dominant","Cop_Veg_2018_ottavi","Cop_Veg_2015_ottavi")] 
dati.cov$dominant <- as.character(dati.cov$dominant)
dati.cov$dominant[is.na(dati.cov$dominant)] <- "Altro"
head(dati.cov)

write.csv2(dati.cov,"Dati-elaborati\\Dati-cov.csv", row.names=FALSE)
