==> 3. Catture - Precipitazioni - Temperatura.ipynb <==
# Creazione matrici Detection Histories delle catture per anno
### Claudia Gabriela Ochoa Martinez (2020)

### Per creare le matrici di Detection histories usiamo il seguente codice, caricando il file excel originale con le informazioni delle nuove catture effettuate. (Se sito e posizionamento trappole rimane invariato)

Daticatt <- read.csv("Dati-elaborati\\Dati-catture.csv", dec=",", sep=";", header = TRUE)
Datioccas <- read.csv("Dati-elaborati\\Dati-occasioni.csv", dec=",", sep=";", header = TRUE)

occasioni_ncols <- length(Datioccas$occasione)
trappole_totali <- max(Datioccas$trappole.posizionate)

# Crea un dataframe vuoto in cui inserire la relazione tra occasioni e trappole

m <- matrix(0, ncol = occasioni_ncols, nrow = trappole_totali)
catture_effettive <- data.frame(m)

# Da un nome ad ogni variabile occasione ed aggiunge
colnames(catture_effettive) <- c(paste0("occ ", 1:occasioni_ncols))

Daticatt_ <- subset(Daticatt,select = c("occasione","sessione","ID_TRAPPOLA","Anno"))
Datioccas$occasione_globale <- seq.int(nrow(Datioccas))
Datioccas_ <- subset(Datioccas, select= c("occasione_globale","occasione","sessione","Anno"))


occasione_cattura <- merge(Datioccas_,Daticatt_)

occasione_cattura <-occasione_cattura[order(occasione_cattura$occasione_globale),]
head(occasione_cattura)

# Riempi la matrice con i valori dicotomici (1 indica almeno una cattura, 0 indica l'assenza di catture)
for (cattura in c(1:nrow(occasione_cattura))) {
    singola_cattura <- occasione_cattura[cattura,]
    trappola <- as.numeric(as.character(singola_cattura$ID_TRAPPOLA))
    occa <- occasione_cattura[cattura,]$occasione_globale
    catture_effettive[trappola, paste0("occ ", occa)] <- 1
}
rownames(catture_effettive) <- c(1:trappole_totali)

head(catture_effettive)

### Orario occasione

orario_occasione_s <- (t(matrix(Datioccas$ALBA..A.T.)))
colnames(orario_occasione_s) <- c(paste0("orario ", 1:occasioni_ncols))
orario_occasione_s[orario_occasione_s == "T"] <- as.character("Al")
orario_occasione_s[orario_occasione_s == "A"] <- as.character("Tr")
orario_occasione <- as.data.frame(rbind(orario_occasione_s[rep(1, trappole_totali), ]))
rownames(orario_occasione) <- c(1:trappole_totali)
tail(orario_occasione)

### Precipitazioni occasione

precipitazioni_occasione_s <- data.frame(t(matrix(Datioccas$Precipitazioni)))
colnames(precipitazioni_occasione_s) <- c(paste0("orario ", 1:occasioni_ncols))
precipitazioni_occasione <- rbind(precipitazioni_occasione_s[rep(1, trappole_totali), ])
rownames(precipitazioni_occasione) <- c(1:trappole_totali)
tail(precipitazioni_occasione)

### Temperatura occasione

temperatura_occasione_s <- data.frame(t(matrix(Datioccas$Temperatura.C)))
colnames(temperatura_occasione_s) <- c(paste0("orario ", 1:occasioni_ncols))
temperatura_occasione <- rbind(temperatura_occasione_s[rep(1, trappole_totali), ])
rownames(temperatura_occasione) <- c(1:trappole_totali)
tail(temperatura_occasione)

velvento_occasione_s <- data.frame(t(matrix(Datioccas$velocita.vento)))
colnames(velvento_occasione_s) <- c(paste0("orario ", 1:occasioni_ncols))
velvento_occasione <- rbind(velvento_occasione_s[rep(1, trappole_totali), ])
rownames(velvento_occasione) <- c(1:trappole_totali)
tail(velvento_occasione)

write.csv2(catture_effettive,"Dati-preparazione-modello\\catture_effettive_totale.csv", row.names = FALSE)
write.csv2(orario_occasione,"Dati-preparazione-modello\\orario_occasione_totale.csv", row.names = FALSE)
write.csv2(precipitazioni_occasione,"Dati-preparazione-modello\\precipitazioni_occasione_totale.csv", row.names = FALSE)
write.csv2(temperatura_occasione,"Dati-preparazione-modello\\temperatura_occasione_totale.csv", row.names = FALSE)
write.csv2(velvento_occasione,"Dati-preparazione-modello\\velvento_occasione_totale.csv", row.names = FALSE)
