==> 7.2 Analisi dei modelli di occupancy creati.ipynb <==
# Analisi Modello Occupancy
### Claudia Gabriela Ochoa Martinez (2020)

library(MuMIn)
library(unmarked)

#### Selezionamo il file generato nello script precedente (6.1. Modelli di Occupancy - Sessione 1 oppure 6.2. Modelli di Occupancy - Sessione 2 )

lista_multimodel <- list.files("Dati-modelli\\", pattern="*.RData", recursive=FALSE, full.names=TRUE)
data.frame(lista_multimodel)

multimodel_scelto <- lista_multimodel[4]
load(multimodel_scelto)
multimodel_scelto

head(multimodel)

#subset(multimodel, logLik == min(logLik))

#### Attraverso la seguente funzione possiamo verificare i modelli migliori. 

get.models(multimodel, 2)

### PASSAGGIO DI MODEL AVERAGING: CONSIDERIAMO CONTEMPORANEAMENTE I MODELLI MIGLIORI CHE DIFFERISCONO TRA LORO
## DI POCO IN TERMINI DI AIC (deltaAIC < 2)
#summary(model.avg(multimodel, subset = delta < 2))

best_model_2018_s2 <- occu(formula = ~MDR + orario + precip + temper + vento + 1 ~ 
    alt_m + dominant + MDR + 1, data = arvicole, control = list(trace = 1, 
    maxit = 1e+07))

summary(best_model_2018_s2)
